Wall of death (there are some issues with destroyed tiles, basically it's not implemented)

+++ greengrappler/core/src/main/java/com/meros/playn/core/UtilMethods.java
@@ -14,7 +14,7 @@ public class UtilMethods {
 			return aMin;
 		if (aRatio > 1.0f)
 			return aMax;
-		return (int) (aMin + (aMax - aMin) * aRatio);
+		return aMin + (aMax - aMin) * aRatio;
 	}
 	
 	public static float2 sincos(double d) {
